Add @mui/material-nextjs dependency and refactor RecipeList component for improved recipe management
