+++ package.json
@@ -13,10 +13,12 @@
 		"cf-typegen": "wrangler types --env-interface CloudflareEnv env.d.ts"
 	},
 	"dependencies": {
+		"@emotion/cache": "^11.14.0",
 		"@emotion/react": "^11.14.0",
 		"@emotion/styled": "^11.14.0",
 		"@mui/icons-material": "^6.4.5",
 		"@mui/material": "^6.4.5",
+		"@mui/material-nextjs": "^6.4.3",
 		"js-cookie": "^3.0.5",
 		"next": "^15.1.7",
 		"react": "^19.0.0",
